[UPDATE] changed the config file format changed the engine logic to download from given index with different sort key words

+++ utils/CSVWriter.go
@@ -13,28 +13,40 @@ func CreateFilePath(folderName string) {
 	if err != nil {
 		log.Fatalln(err)
 	}
-	err = os.Mkdir(path+"/Result/"+folderName, 0755)
+	exists, err := exists(path + "/Result/" + folderName)
 	if err != nil {
-		log.Println(err)
+		log.Fatalln(err)
+	}
+
+	if !exists {
+		err = os.Mkdir(path+"/Result/"+folderName, 0755)
+		if err != nil {
+			log.Println(err)
+		}
 	}
 	FolderPath = path + "/Result/" + folderName
 }
 
 func InitCSVWriter(folderName string) error {
-
-	file, err := os.Create(FolderPath + "/details.csv")
+	exists, err := exists(FolderPath + "/details.csv")
 	if err != nil {
-		return err
+		log.Fatalln(err)
 	}
-	defer file.Close()
+	if !exists {
+		file, err := os.Create(FolderPath + "/details.csv")
+		if err != nil {
+			return err
+		}
+		defer file.Close()
 
-	initColumns := []string{"id", "name", "description", "category", "tags", "rating", "taken_at", "shutter_speed", "focal_length", "aperture", "camera", "Lens", "iso", "location", "latitude", "longitude", "nsfw"}
-	csvwriter := csv.NewWriter(file)
-	err = csvwriter.Write(initColumns)
-	if err != nil {
-		return err
+		initColumns := []string{"id", "name", "description", "category", "tags", "rating", "taken_at", "shutter_speed", "focal_length", "aperture", "camera", "Lens", "iso", "location", "latitude", "longitude", "nsfw"}
+		csvwriter := csv.NewWriter(file)
+		err = csvwriter.Write(initColumns)
+		if err != nil {
+			return err
+		}
+		csvwriter.Flush()
 	}
-	csvwriter.Flush()
 	return nil
 }
 
@@ -51,4 +63,15 @@ func WritePhotoDetailRecord(detail []string) error {
 	}
 	csvwriter.Flush()
 	return nil
+}
+
+func exists(path string) (bool, error) {
+	_, err := os.Stat(path)
+	if err == nil {
+		return true, nil
+	}
+	if os.IsNotExist(err) {
+		return false, nil
+	}
+	return false, err
 }
